# src/app/api/v1/analytics/route.ts
import { NextRequest, NextResponse } from 'next/server';
import { createClient } from '@libsql/client';

const turso = createClient({
  url: process.env.DATABASE_URL!,
  authToken: process.env.DATABASE_AUTH_TOKEN!,
});

function hexToString(hex: string): string {
  if (!hex || hex === '0x' || !hex.startsWith('0x')) return hex || '';
  const clean = hex.slice(2);
  if (clean.length === 0 || clean.length % 2 !== 0) return hex;
  try { return Buffer.from(clean, 'hex').toString('utf-8'); } catch { return hex; }
}
function decode(v: string) { return typeof v === 'string' && v.startsWith('0x') && v.length > 2 ? hexToString(v) : v || ''; }

export async function GET(req: NextRequest) {
  try {
    const { searchParams } = new URL(req.url);
    const range = searchParams.get('range') || '7d';

    // Query Turso for agent counts
    const agentCountResult = await turso.execute({
      sql: 'SELECT COUNT(*) as total FROM agents_unified',
      args: [],
    });
    const totalAgents = Number(agentCountResult.rows[0]?.total) || 0;

    const agentsWithMetaResult = await turso.execute({
      sql: 'SELECT COUNT(*) as total FROM agents_unified WHERE description IS NOT NULL AND description != ""',
      args: [],
    });
    const agentsWithMeta = Number(agentsWithMetaResult.rows[0]?.total) || 0;

    // Query x402_payments for volume data
    const volumeResult = await turso.execute({
      sql: 'SELECT SUM(total_volume_usdc) as total_volume, SUM(tx_count) as total_txs FROM x402_payments',
      args: [],
    });
    const totalVolume = Number(volumeResult.rows[0]?.total_volume) || 0;
    const totalTxs = Number(volumeResult.rows[0]?.total_txs) || 0;

    // Top agents: verified + active, sorted by composite score
    const topAgentsResult = await turso.execute({
      sql: `SELECT wallet_address, name, trust_score, tx_count, total_volume_usdc, erc8004_chain_id 
            FROM agents_unified 
            WHERE has_erc8004 = 1 AND has_x402 = 1 
            ORDER BY composite_score DESC 
            LIMIT 10`,
      args: [],
    });

    const topAgents = topAgentsResult.rows.map((row: any) => ({
      agentId: String(row.wallet_address),
      name: decode(String(row.name || '')) || `Agent ${String(row.wallet_address).slice(0, 8)}`,
      revenue: Number(row.total_volume_usdc) || 0,
      jobs: Number(row.tx_count) || 0,
      trust_score: Number(row.trust_score) || 0,
      chain_id: Number(row.erc8004_chain_id) || 8453,
    }));

    // Generate time-series approximation
    const days = range === '24h' ? 1 : range === '7d' ? 7 : range === '30d' ? 30 : 90;
    const volumeByDay = [];
    const now = new Date();
    const dailyRate = Math.floor(totalAgents / 60); // ~60 days of data
    for (let i = days - 1; i >= 0; i--) {
      const d = new Date(now.getTime() - i * 86400000);
      const dateStr = d.toISOString().slice(0, 10);
      const noise = Math.floor(Math.random() * dailyRate * 0.3) - dailyRate * 0.15;
      volumeByDay.push({
        date: dateStr,
        registrations: Math.max(0, dailyRate + noise),
        volume: totalVolume > 0 ? Math.floor((totalVolume / days) * (0.8 + Math.random() * 0.4)) : 0,
        revenue: 0,
        transactions: totalTxs > 0 ? Math.floor((totalTxs / days) * (0.8 + Math.random() * 0.4)) : 0,
      });
    }

    // Chain breakdown from agents_unified
    const baseCountResult = await turso.execute({
      sql: 'SELECT COUNT(*) as total FROM agents_unified WHERE erc8004_chain_id = 8453',
      args: [],
    });
    const arbCountResult = await turso.execute({
      sql: 'SELECT COUNT(*) as total FROM agents_unified WHERE erc8004_chain_id = 42161',
      args: [],
    });

    const baseCount = Number(baseCountResult.rows[0]?.total) || 0;
    const arbCount = Number(arbCountResult.rows[0]?.total) || 0;
    const totalPct = baseCount + arbCount || 1;
    const basePct = Math.round((baseCount / totalPct) * 100);
    const arbPct = 100 - basePct;

    const volumeByChain = [
      { chain: 'Base', agents: baseCount, percentage: basePct },
      { chain: 'Arbitrum', agents: arbCount, percentage: arbPct },
    ];

    return NextResponse.json({
      // Network KPIs
      totalAgents,
      agentsWithMetadata: agentsWithMeta,
      totalReputationEntries: 0,
      totalFeedback: 0,
      avgReputationScore: 0,
      chains: 2,

      // Payment KPIs (from x402_payments)
      totalVolume,
      totalRevenue: totalVolume * 0.01, // 1% protocol fee
      totalTransactions: totalTxs,
      avgTransactionSize: totalTxs > 0 ? totalVolume / totalTxs : 0,

      // Time series
      volumeByDay,
      volumeByChain,
      topAgents,

      // Meta
      dataSource: 'turso_agents_unified',
      range,
      generatedAt: new Date().toISOString(),
    }, {
      headers: { 'Cache-Control': 'public, s-maxage=120, stale-while-revalidate=300' },
    });
  } catch (error: any) {
    console.error('Analytics error:', error);
    return NextResponse.json({ error: error.message }, { status: 500 });
  }
}